@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-  	<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+  	<TargetFrameworks>net6.0</TargetFrameworks>
 	<Nullable>enable</Nullable>
   	<RootNamespace>CosmosQuery</RootNamespace>
   	<PackageId>CosmosQuery</PackageId>
